requirements.txt
Django==4.0.3
djangorestframework==3.13.1
Pillow==9.0.1
drf-spectacular==0.21.2
djangorestframework-simplejwt==5.1.0